+++ Group11_p4/Phase1/Code/ground_truth_TUM.py
