+++ Group11_p4/Phase1/Code/imgs/euroc_mh_01_easy.png
