+++ Group11_p4/Phase1/Output/traj_gt.txt
